🧹 Tidy: LandingContent.tsxのリファクタリング
