+++ .jules/tidy.md
@@ -186,3 +186,6 @@
 
 アクション:
 - まだ `Growth` や `Sleep` に同様の重複が残っている可能性があるため、順次この「正規化アダプターパターン」を適用していく。
+2026-02-21 - [Fat Componentの分割] LandingContent.tsxのリファクタリング
+学び: 巨大なランディングページのコンポーネントは、セクションごとに分割することで可読性が劇的に向上する。静的なコンテンツが多い場合、各セクションを独立したコンポーネントとして切り出すことで、親コンポーネントが単なるレイアウト構成要素となり、保守性が高まる。
+アクション: ページのメインコンポーネントが500行を超えた場合、セクション単位での分割を積極的に検討する。
